Use password in OpenVPN management interface
This avoids a warning being emitted by `openvpn`. Interestingly, the use
of password is backwards from what I would expect. The local management
server gives the password back to `openvpn`, not the other way around.
In this way, it does not seem to improve the security of potential
management interface connections.

Regardless, the management server will still only ever accept a single
connection before it stops listening. This should avoid later
connections discovering it and trying to be malicious.

This also adds additional debug logging around the management client to
help surface messages it receives and when errors occur.

https://github.com/dpb587/ssoca/issues/14