openvpn: Add automatic workaround for openssl/1.1.1
This tries to detect if openssl >= 1.1.1 is being used and automatically
disable the preferred management interface method for managing credentials.
The openvpn process currently errors unless --static-certificates is used
due to library changes in how it processes TLS handshakes and the padding
modes it uses.

Hopefully a patch in openvpn and related management API change will be
released which future versions can rely on instead.

https://community.openvpn.net/openvpn/ticket/1216
https://github.com/dpb587/ssoca/issues/13